UserInfo de EivAS

+++ webapp/src/main/java/com/cairone/example/webapp/oauth2/EivasOAuth2UserInfo.java
@@ -10,7 +10,7 @@ public class EivasOAuth2UserInfo extends OAuth2UserInfo {
 
     @Override
     public String getId() {
-        return ((Integer) attributes.get("id")).toString();
+        return attributes.get("id").toString();
     }
 
     @Override
